combined accountinfo and ethers context into one global wallet provider
