+++ components/Layout.tsx
@@ -1,8 +1,6 @@
 import React, { ReactNode } from 'react'
 import Head from 'next/head'
-import AccountInfoProvider from '../context/AccountInfoContext'
-import EthersProvider from '../context/EthersContext'
-
+import WalletProvider from '../context/WalletContext'
 interface Props {
     children?: ReactNode
 }
@@ -13,15 +11,13 @@ const style = {
 
 const Layout = ({ children }: Props) => {
     return (
-        <EthersProvider>
-            <AccountInfoProvider>
-                <Head>
-                    <title>GammaSwap</title>
-                    <link rel="icon" href="/favicon.ico" />
-                </Head>
-                <main className={style.wrapper}>{children}</main>
-            </AccountInfoProvider>
-        </EthersProvider>
+        <WalletProvider>
+            <Head>
+                <title>GammaSwap</title>
+                <link rel="icon" href="/favicon.ico" />
+            </Head>
+            <main className={style.wrapper}>{children}</main>
+        </WalletProvider>
     )
 }
